fix(management): redirect to app default tab on unknown tab param
WzManagementMain renders blank when tab is missing or set to an
unsupported value, since its Switch has no fallback. Add a Redirect
whose target is derived from the currently active app (Endpoint
Groups, Status, Cluster, Statistics, Logs, or Settings), mirroring the
Settings fix. Extract the shared "resolve default tab from the current
app" logic into applications.ts (getCurrentAppDefaultTabSearch) and
reuse it from Settings instead of duplicating it.
